+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.21.15][] - 2020-04-01
+
+### Added
+
+-   Added `contentRef` prop to `Dialog` component
+-   Added `linkRef` on `Link` component
+
+### Fixed
+
+-   Fixed blurry Tooltip
+
 ## [0.21.14][] - 2020-03-27
 
 ### Added
@@ -415,3 +426,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.21.13]: https://github.com/lumapps/design-system/tree/v0.21.13
 [unreleased]: https://github.com/lumapps/design-system/compare/v0.21.14...HEAD
 [0.21.14]: https://github.com/lumapps/design-system/tree/v0.21.14
+[unreleased]: https://github.com/lumapps/design-system/compare/v0.21.15...HEAD
+[0.21.15]: https://github.com/lumapps/design-system/tree/v0.21.15